app/lib/db/config.ts
import { drizzle } from 'drizzle-orm/pglite';
// Import will be used when implementing migrations
// eslint-disable-next-line @typescript-eslint/no-unused-vars
import { migrate } from 'drizzle-orm/pglite/migrator';
import * as schema from './schema';

// This will be initialized when we integrate pglite
let db: ReturnType<typeof drizzle> | null = null;

export async function initDatabase() {
  if (db) return db;

  // TODO: Initialize pglite database
  // const pglite = await createPglite();
  // await pglite.start();
  // db = drizzle(pglite, { schema });
  
  // Run migrations
  // await migrate(db, { migrationsFolder: './migrations' });
  
  return db;
}

export async function getDatabase() {
  if (!db) {
    db = await initDatabase();
  }
  return db;
}

// Template embedding functions to be implemented
export async function storeTemplateEmbedding(
  // eslint-disable-next-line @typescript-eslint/no-unused-vars
  path: string,
  // eslint-disable-next-line @typescript-eslint/no-unused-vars
  embedding: number[],
  // eslint-disable-next-line @typescript-eslint/no-unused-vars
  metadata: Omit<schema.TemplateEmbedding, 'id' | 'embedding' | 'createdAt' | 'updatedAt'>
) {
  const database = await getDatabase();
  if (!database) return null;

  // TODO: Implement when pglite is integrated
  return null;
}

export async function findSimilarTemplates(
  // eslint-disable-next-line @typescript-eslint/no-unused-vars
  embedding: number[],
  // eslint-disable-next-line @typescript-eslint/no-unused-vars
  limit = 5
): Promise<Array<schema.TemplateEmbedding & { similarity: number }>> {
  const database = await getDatabase();
  if (!database) return [];

  // TODO: Implement vector similarity search using pgvector
  return [];
}

// Usage tracking functions
export async function trackTemplateUsage(
  // eslint-disable-next-line @typescript-eslint/no-unused-vars
  templateId: string,
  // eslint-disable-next-line @typescript-eslint/no-unused-vars
  prompt: string,
  // eslint-disable-next-line @typescript-eslint/no-unused-vars
  similarity: number,
  // eslint-disable-next-line @typescript-eslint/no-unused-vars
  successful: boolean
) {
  const database = await getDatabase();
  if (!database) return;

  // TODO: Implement usage tracking
}

export async function recordTemplateFeedback(
  // eslint-disable-next-line @typescript-eslint/no-unused-vars
  templateId: string,
  // eslint-disable-next-line @typescript-eslint/no-unused-vars
  prompt: string,
  // eslint-disable-next-line @typescript-eslint/no-unused-vars
  rating: number,
  // eslint-disable-next-line @typescript-eslint/no-unused-vars
  comment?: string
) {
  const database = await getDatabase();
  if (!database) return;

  // TODO: Implement feedback recording
}
